requirements.txt
# Dependencias para conexión a SQL Server
pyodbc==4.0.39
sqlalchemy==2.0.23

# Dependencias para RabbitMQ
pika==1.3.2

# Dependencias para scraping
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
lxml==4.9.3

# Dependencias adicionales
python-dotenv==1.0.0
schedule==1.2.0
logging==0.4.9.6
